restaurant/requirements.txt
grpcio
grpcio-tools
requests
pymongo